+++ src/Component/OpenApi2/Tests/fixtures/issue-770/expected/Normalizer/IndoorMapIndoorMapSummaryNormalizer.php
@@ -38,13 +38,13 @@ class IndoorMapIndoorMapSummaryNormalizer implements DenormalizerInterface, Norm
             return new Reference($data['$recursiveRef'], $context['document-origin']);
         }
         if (\array_key_exists('apCount', $data) && \is_int($data['apCount'])) {
-            $data['apCount'] = (double) $data['apCount'];
+            $data['apCount'] = (float) $data['apCount'];
         }
         if (\array_key_exists('latitude', $data) && \is_int($data['latitude'])) {
-            $data['latitude'] = (double) $data['latitude'];
+            $data['latitude'] = (float) $data['latitude'];
         }
         if (\array_key_exists('longitude', $data) && \is_int($data['longitude'])) {
-            $data['longitude'] = (double) $data['longitude'];
+            $data['longitude'] = (float) $data['longitude'];
         }
         if (\array_key_exists('id', $data)) {
             $object->setId($data['id']);
