+++ src/Component/OpenApi3/Tests/fixtures/issue-337/expected/Normalizer/GbCompanyReportExampleResponseReportFinancialStatementsItemOtherFinancialsNormalizer.php
@@ -38,10 +38,10 @@ class GbCompanyReportExampleResponseReportFinancialStatementsItemOtherFinancials
             return new Reference($data['$recursiveRef'], $context['document-origin']);
         }
         if (\array_key_exists('workingCapital', $data) && \is_int($data['workingCapital'])) {
-            $data['workingCapital'] = (double) $data['workingCapital'];
+            $data['workingCapital'] = (float) $data['workingCapital'];
         }
         if (\array_key_exists('netWorth', $data) && \is_int($data['netWorth'])) {
-            $data['netWorth'] = (double) $data['netWorth'];
+            $data['netWorth'] = (float) $data['netWorth'];
         }
         if (\array_key_exists('contingentLiabilities', $data)) {
             $object->setContingentLiabilities($data['contingentLiabilities']);
